requirements.txt
joblib==1.0.1
kafka-python==2.0.2
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
threadpoolctl==2.2.0
